fix(api): fix failing auth
The exact cause of the issue is still unknown to me.
- Sometimes the `flow.fetch_token` was failing with `invalid_grant` error which is a
vague error that google returns if there's something wrong with the
request.
- Sometimes the `id_token.verify_oauth2_token` fails due to system clock
being in unsync state with the server's clock.
  - Added 5 seconds as `clock_skew_in_seconds` to relax this difference
to 5 seconds. If later an error like: `Token used too early, 1658076534
< 1658076536. Check that your computer's clock is set correctly.`
arises, we probably need to increase this limit accordingly.